Small Update [Spirit + Tankman]
[Spirit]
Set Idle offset to 0, matching the rest of the cast.
Adjusted the other singing offsets  to match the idle. (All are now closer to [0, 0], except for singDOWN due to it's first frame)
Moved singLEFT down & singRIGHT left by one pixel to better match each other's plus the idle's faces. (Character pixel not position)
Moved singDOWN a bit lower to differentiate it more from the idle.
Moved the character a bit up, so Spirit is floating.
Added a small camera offset so the evil tree is visible in base "throns".

[Tankman]
Moved singRIGHT a bit to the left to stop it from sliding so much. Also making the bottom of the right foot more consistent.